requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.23.2
streamlit>=1.28.1
pydantic>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
setuptools>=68.0.0
wheel>=0.41.0

# Data processing and ML
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# LangChain
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0

# Evaluation and visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
sqlite-utils>=3.0